aider: Add fields for authorization to the `application` struct.
# Aider chat conversation:

USER: add fields for authorization to the application struct. they should be able to pass auth header and auth key for now, but it should support basic auth too

ASSISTANT: To add fields for authorization to the `application` struct, we need to modify the `application` struct in the `main.go` file. We will add `AuthHeader` and `AuthKey` fields for passing the authorization header and key. We will also add `BasicAuthUsername` and `BasicAuthPassword` fields for basic authentication.

Here is the change:

```go
main.go
<<<<<<< SEARCH
type application struct {
	Name        string
	URL         string
	Description string
	httpResp    httpResp
}
=======
type application struct {
	Name             string
	URL              string
	Description      string
	httpResp         httpResp
	AuthHeader       string
	AuthKey          string
	BasicAuthUsername string
	BasicAuthPassword string
}
>>>>>>> REPLACE
```

This change adds the `AuthHeader`, `AuthKey`, `BasicAuthUsername`, and `BasicAuthPassword` fields to the `application` struct. These fields can be used to store the necessary information for authorization.