aider: Update `checkServers` function to use `AuthHeader` and `AuthKey` fields if they are not empty when making a request to the app.URL.
# Aider chat conversation:

USER: update the `checkServers` function to use the `AuthHeader` and `AuthKey` fields if they are not empty when making a request to the app.URL

ASSISTANT: To update the `checkServers` function to use the `AuthHeader` and `AuthKey` fields when making a request, we need to modify the `checkServers` function in the `main.go` file. We will add a condition to check if the `AuthHeader` and `AuthKey` fields are not empty, and if so, we will add them to the request header.

Here is the change:

```go
main.go
<<<<<<< SEARCH
func (m model) checkServers(d time.Duration) tea.Cmd {
	return tea.Tick(d, func(t time.Time) tea.Msg {
		msg := make(statusMsg)

		for _, app := range m.applications {
			res, err := m.client.Get(app.URL)
			if err != nil {
				msg[app.URL] = 0
				continue
			}
			msg[app.URL] = res.StatusCode
		}
		return msg
	})
}
=======
func (m model) checkServers(d time.Duration) tea.Cmd {
	return tea.Tick(d, func(t time.Time) tea.Msg {
		msg := make(statusMsg)

		for _, app := range m.applications {
			req, err := http.NewRequest("GET", app.URL, nil)
			if err != nil {
				msg[app.URL] = 0
				continue
			}

			if app.AuthHeader != "" && app.AuthKey != "" {
				req.Header.Add(app.AuthHeader, app.AuthKey)
			}

			res, err := m.client.Do(req)
			if err != nil {
				msg[app.URL] = 0
				continue
			}
			msg[app.URL] = res.StatusCode
		}
		return msg
	})
}
>>>>>>> REPLACE
```

This change modifies the `checkServers` function to create a new HTTP request using `http.NewRequest` instead of directly calling `m.client.Get`. This allows us to modify the request headers before sending the request. If the `AuthHeader` and `AuthKey` fields are not empty, they are added to the request header using `req.Header.Add`. The request is then sent using `m.client.Do(req)`.