aider: Added TOML struct tags to the `application` struct for the public fields.
# Aider chat conversation:

USER: add toml struct tags to the `application` struct for the public fields

ASSISTANT: To add TOML struct tags to the `application` struct for the public fields, we need to modify the `main.go` file.

Here is the *SEARCH/REPLACE block* for the changes:

```go
main.go
<<<<<<< SEARCH
type application struct {
	Name              string
	URL               string
	Description       string
	httpResp          httpResp
	AuthHeader        string
	AuthKey           string
	BasicAuthUsername string
	BasicAuthPassword string
}
=======
type application struct {
	Name              string `toml:"name"`
	URL               string `toml:"url"`
	Description       string `toml:"description"`
	httpResp          httpResp
	AuthHeader        string `toml:"authHeader"`
	AuthKey           string `toml:"authKey"`
	BasicAuthUsername string `toml:"basicAuthUsername"`
	BasicAuthPassword string `toml:"basicAuthPassword"`
}
>>>>>>> REPLACE
```

This change adds TOML struct tags to the public fields of the `application` struct. These tags will be used when decoding the TOML configuration file.