aider: Updated the `View` method to output the `piHoleSummary` as a table using the `github.com/charmbracelet/bubbles/table` package.
# Aider chat conversation:

USER: using "github.com/charmbracelet/bubbles/table", in the `View` method output the `piHoleSummary` as a table

ASSISTANT: To output the `piHoleSummary` as a table using the `github.com/charmbracelet/bubbles/table` package, we will need to make changes to the `model.go` file. We will add the table to the `View` method when the `showPiHoleDetail` flag is true.

First, ensure that the `github.com/charmbracelet/bubbles/table` package is imported and used in the `model.go` file. If it's not already in the `go.mod` file, you will need to add it by running `go get github.com/charmbracelet/bubbles/table`.

Here are the steps we will follow to make the changes:

1. Import the `table` package in `model.go`.
2. Create a new table model in the `model` struct to hold the Pi-hole summary table.
3. Initialize the table model in the `Init` method of the `model` struct.
4. Update the `View` method to render the Pi-hole summary as a table when `showPiHoleDetail` is true.

Let's start with the import statement and the addition of the table model to the `model` struct:

```go
model.go
<<<<<<< SEARCH
import (
	"fmt"
	"log"
	"net/http"
	"strings"
	"time"

	"github.com/charmbracelet/bubbles/list"
	tea "github.com/charmbracelet/bubbletea"
	"github.com/charmbracelet/lipgloss"
)
=======
import (
	"fmt"
	"log"
	"net/http"
	"strings"
	"time"

	"github.com/charmbracelet/bubbles/list"
	"github.com/charmbracelet/bubbles/table"
	tea "github.com/charmbracelet/bubbletea"
	"github.com/charmbracelet/lipgloss"
)

// model is the bubbletea model
type model struct {
	// ... other fields ...

	piHoleTable table.Model
}
>>>>>>> REPLACE
```

Next, we will initialize the table model in the `Init` method:

```go
model.go
<<<<<<< SEARCH
func (m model) Init() tea.Cmd {
	return m.checkApplications(10 * time.Millisecond)
}
=======
func (m model) Init() tea.Cmd {
	m.piHoleTable = table.New()
	m.piHoleTable.SetHeader([]string{"Metric", "Value"})
	m.piHoleTable.SetWidth(100)
	m.piHoleTable.SetStyle(table.StyleLight)
	m.piHoleTable.BorderStyle(lipgloss.NormalBorder())
	return m.checkApplications(10 * time.Millisecond)
}
>>>>>>> REPLACE
```

Finally, we will update the `View` method to render the Pi-hole summary as a table:

```go
model.go
<<<<<<< SEARCH
	if m.showPiHoleDetail {
		// Render the detailed page for Pi-hole with actual statistics
		var piHoleDetailBuilder strings.Builder
		piHoleDetailBuilder.WriteString(detailHeaderStyle.Render("Pi-hole Detailed View") + "\n\n")
		piHoleDetailBuilder.WriteString(fmt.Sprintf(detailHeaderStyle.Render("Total Queries: ") + detailDataStyle.Render(m.piHoleSummary.DNSQueries) + "\n"))
		piHoleDetailBuilder.WriteString(fmt.Sprintf(detailHeaderStyle.Render("Queries Blocked: ") + detailDataStyle.Render(m.piHoleSummary.AdsBlocked) + "\n"))
		piHoleDetailBuilder.WriteString(fmt.Sprintf(detailHeaderStyle.Render("Percentage Blocked: ") + detailDataStyle.Render(m.piHoleSummary.AdsPercentage+"%%") + "\n"))
		piHoleDetailBuilder.WriteString(fmt.Sprintf(detailHeaderStyle.Render("Domains on Adlist: ") + detailDataStyle.Render(m.piHoleSummary.DomainsBlocked) + "\n"))
		return piHoleDetailBuilder.String()
	}
=======
	if m.showPiHoleDetail {
		// Update the table with Pi-hole statistics
		m.piHoleTable.SetRows([]table.Row{
			{"Total Queries", m.piHoleSummary.DNSQueries},
			{"Queries Blocked", m.piHoleSummary.AdsBlocked},
			{"Percentage Blocked", m.piHoleSummary.AdsPercentage + "%"},
			{"Domains on Adlist", m.piHoleSummary.DomainsBlocked},
		})

		// Render the table
		return detailHeaderStyle.Render("Pi-hole Detailed View") + "\n\n" + m.piHoleTable.View()
	}
>>>>>>> REPLACE
```

Please note that you will need to run `go get github.com/charmbracelet/bubbles/table` to add the table package to your `go.mod` file if it's not already there. If you need assistance with that or any other changes, let me know!